+++ CHANGELOG.md
@@ -7,10 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 
+## [1.1.0] - 2020-03-23
+### Changed
+* publicMarketData.bitcoinAverage error handling support By [@Rincorpes](https://github.com/Rincorpes)
+
+
 ## [1.0.11] - 2020-02-23
 ### Changed
 * CONTRIBUTING.md By [@Rincorpes](https://github.com/Rincorpes)
-  * Pull Request Process resction improvement
+  * Pull Request Process section improvement
 
 
 ## [1.0.10] - 2020-02-23